feat: Use Naismith helper during route calculation and route saving for ETAs
